+++ src/meltano/core/plugin/base.py
@@ -1,4 +1,6 @@
 import yaml
+import fnmatch
+from collections import namedtuple
 from enum import Enum
 
 
@@ -25,21 +27,36 @@ class PluginType(YAMLEnum):
 
 class Plugin:
     def __init__(
-        self, plugin_type: PluginType, name: str, pip_url=None, config=None, **extras
+        self,
+        plugin_type: PluginType,
+        name: str,
+        pip_url=None,
+        config=None,
+        select=None,
+        **extras
     ):
         self.name = name
         self.type = plugin_type
         self.pip_url = pip_url
         self.config = config
+        self._select = set(select or [])
         self._extras = extras or {}
 
     def canonical(self):
-        return {
+        canonical = {
             "name": self.name,
             "pip_url": self.pip_url,
             "config": self.config,
             **self._extras,
         }
+
+        if self._select:
+            canonical.update({"select": list(self._select)})
+
+        if self._extras:
+            canonical.update(**self._extras)
+
+        return canonical
 
     @property
     def config_files(self):
@@ -49,6 +66,17 @@ class Plugin:
     @property
     def output_files(self):
         return []
+
+    @property
+    def select(self):
+        return self._select or {"*.*"}
+
+    @select.setter
+    def select(self, patterns):
+        self._select = set(patterns)
+
+    def add_select_filter(self, filter: str):
+        self._select.add(filter)
 
     def __eq__(self, other):
         return self.name == other.name and self.type == other.type